refactor: extracted util + parser into their own files, small refactor in parser
